+++ application.c
@@ -116,7 +116,8 @@ err:
 }
 
 bool
-application_execute(const struct application *app, const struct prompt *prompt, const char *launch_prefix)
+application_execute(const struct application *app, const struct prompt *prompt,
+                    const char *launch_prefix, const char *xdg_activation_token)
 {
     const char32_t *ptext = prompt_text(prompt);
     const size_t c32len = c32tombs(NULL, ptext, 0);
@@ -214,6 +215,13 @@ application_execute(const struct application *app, const struct prompt *prompt, 
         if (path != NULL) {
             if (chdir(path) == -1)
                 LOG_ERRNO("failed to chdir to %s", path);
+        }
+
+        if (xdg_activation_token != NULL) {
+            /* Prepare client for xdg activation startup */
+            setenv("XDG_ACTIVATION_TOKEN", xdg_activation_token, 1);
+            /* And X11 startup notifications */
+            setenv("DESKTOP_STARTUP_ID", xdg_activation_token, 1);
         }
 
         /* Redirect stdin/stdout/stderr -> /dev/null */